Adds a 'pay what you want' option for donation
