+++ package.json
@@ -88,6 +88,7 @@
     "react-chartjs-2": "^2.11.1",
     "react-cookie": "^4.0.3",
     "react-cookie-consent": "^6.4.1",
+    "react-currency-input-field": "^3.6.0",
     "react-datepicker": "^3.6.0",
     "react-dates": "^21.8.0",
     "react-dom": "^17.0.2",
